Prevent XSSI access to `MODx.config` by requiring auth token (#15644)
Reported by Solar Security in security report 3837, it's possible to bypass the Same Origin Policy when including the modx.config.js.php connector as a script tag. If a logged-in user is tricked into visiting a malicious site, that would allow the attacker to access information meant for the manager typically available in MODx.config. This information is already filtered to exclude sensitive data like passwords, but may still leak the rest of the MODX configuration including extras.

To avoid this XSSI vulnerability, the connector is changed to require permissions the same way as other processors and the manager controller injects the HTTP_MODAUTH URL parameter. Attackers will now need first-party access to the auth token to abuse.

To avoid further leaking the auth token, it's only set temporarily while parsing the header.tpl. Should an attacker manage to inject arbitrary smarty tags, that does not provide them the token.

Also made sure an error message is returned when the user does not have permission.

Custom manager themes that override the header.tpl may require the same fix or break.