Prevent access to hashed user password, cachepwd, salt, and sessionid [SEC-3837]
This addresses a variety of different ways a manager user with permission to manage users might be able to access hashed passwords, the cachepwd, or a users' last sessionid. Follows from report by Solar Security who reported it was included in the security/group/user/getList response, after which I did some more searching and found a few more places that could return some of these.

This is not super interesting from a security perspective, as any user that has permission to manage users could change their password and just take over an account that way. But as these fields are not supposed to be readable, it's good to lock 'm down anyway.