Pass auth token to Browser controller config (#15692)
After #15644, the modx config is only available when an authentication token is provided in the parameters. The ?a=browser controller is a special one used by extras (in particular tinymce rte, probably tinymce as well) that provices the regular media browser (?a=media/browser) without the manager frame. It's meant to be opened in an iframe.

This PR adds the necessary token (which is already being set in the BrowserIndexController) to the request to restore the functionality of third party editors.