Add modRest->request method for easier dynamic request types
modRestClient and other request handlers out there (eg Guzzle) offer a generic request method that can be provided the method to use; this was lacking in modRest. By adding it we can keep more BC in package management and don't have to resort to if/else/switch constructs to accept the method to use.