Catch fatal errors/exceptions during package installation nicely instead of breaking
If a package installation triggers a fatal error or exception, currently it prevents the installation from halting normally (i.e. the console window updating with the failure) and it also doesn't show the error. That requires users to dig through error logs or the raw ajax request to find the error. With this patch, the error message is actually automatically added to the console immediately.

This also opens up the ability for packages (resolves/validators) to throw an exception to indicate a failure.