Assorted modRest improvements (#15777)
* Fix XML parsing issue when multiple children types exist

Given an XML structure like this (from package management):

home
- name
- topdownloaded<id,name>
- topdownloaded<id,name>
- ...
- newest<id,name>
- newest<id,name>
- ...

it would create an array structure like this:

[
  name
  topdownloaded => [
     [id,name]
     [id,name]
   ]
   newest => [
     id
     name
     [id,name]
   ]
]

because the $first detection only ran on the first array-type it encountered, and not the second. With this fix it's correct:

[
  topdownloaded => [
     [id,name]
     [id,name]
   ]
   newest => [
     [id,name]
     [id,name]
   ]
]

* Add modRest->request method for easier dynamic request types

modRestClient and other request handlers out there (eg Guzzle) offer a generic request method that can be provided the method to use; this was lacking in modRest. By adding it we can keep more BC in package management and don't have to resort to if/else/switch constructs to accept the method to use.

* Catch fatal errors/exceptions during package installation nicely instead of breaking

If a package installation triggers a fatal error or exception, currently it prevents the installation from halting normally (i.e. the console window updating with the failure) and it also doesn't show the error. That requires users to dig through error logs or the raw ajax request to find the error. With this patch, the error message is actually automatically added to the console immediately.

This also opens up the ability for packages (resolves/validators) to throw an exception to indicate a failure.

* Fix duplicate slash being added to the URL if it doesn't already start with a slash, but the baseUrl ends with one

* Fix inability to read CDATA elements

* Fix lexicon string

Co-authored-by: Jason Coward <jason@opengeek.com>