Fix creating/updating absolute static elements causing a copy to appear relative to MODX
The source being loaded, regardless of the static_source option, causes an in-memory representation of the element to have a source assigned despite it being set to "(none)". That source then calls an error when it's not needed.

As for isStaticSourceMutable; the modElement::getSourceFile method already handles checking readability and returns either a bool, an absolute url, or a source-relative url. Media sources don't have an isWritable method so for now that's just being dropped. The createObject() method has the side effect of creating files which seems unexpected.