Fix and improve GetAreas
Fix issue with a non-named area being blank in the list. Also fine-tuned the building of the list array so that duplicate areas show only one list entry along with the correct count of setting entries. Lastly, fixed the sorting of the list to make it based on the label and not the actual value.