Refactor the input filter parsing of commands and modifiers
Rather than relying on regex, which is super complex and doesn't handle nesting well, this parses the string one character at a time to really understand what the implications of a certain nested tag are.

This makes sure that any odd ` or : in a modifier doesn't affect the root tag that is being processed.

Also expanded the test suite to confirm the expected behavior.

Related to:

- #13250: one of the earliest reports of issues with nested tags, with lots of discussions.
- #14095: report of nested conditionals causing broken tags
- #16254: proposed revert of #14458 which introduced issues with nested output filters (` characters)