Grid filtering via URL parameters -- Usergroup ACL (#16089)
* Add remembering filter properties of template TVs grid in browser url

* Fix and improve GetAreas

Fix issue with a non-named area being blank in the list. Also fine-tuned the building of the list array so that duplicate areas show only one list entry along with the correct count of setting entries. Lastly, fixed the sorting of the list to make it based on the label and not the actual value.

* Required method additions for new persistence feature

* Application of persistence feature to ACL User Group panel

* A few pre-submission tweaks

* grunt build

Co-authored-by: Ruslan-Aleev <aleev.insmile@gmail.com>
Co-authored-by: Jason Coward <jason@opengeek.com>