Add IDs to toolbar elements
If toolbar elements i.e. have to be hidden in a class that extends MODx.grid.SettingsGrid, Ext.getCmp needs an ID to access the toolbar element.

In #16089 the ID was removed. This PR fixes that but with an ID that is set variable with the Ext.id of the current MODx.grid.SettingsGrid. So multiple MODx.grid.SettingsGrid won't interfere with each other because of using the same ID multiple.