Fix Trash Manager infinite loading error (#16415)
### What is he doing?
Fixes the "infinite list loading" error.


### Where did this error come from?
https://github.com/modxcms/revolution/pull/16228
This change causes an error when starting the processor
"resource/trash/getlist".


### How to test
Option 1: Delete the document and go to the trash. Then clean it. There
will be an "infinite loading" of the list due to an error. Option 2: Go
to the shopping cart and click on the "Update list" icon. There will be
an "infinite loading" of the list due to an error.


### Decision
It is necessary to return the sorting so that the xPDOQuery object is
returned