Update Template Build Process (#16540)
### What does it do?
Replaces most legacy dependencies with current versions—with the
exception of bourbon, neat, and fontawesome—and drops others that are no
longer relevant (such as imageoptim).

### Why is it needed?
Bring 2.x more in line with 3.x, mainly allow use of modern js features.

### How to test

1. Run the rebuild processes, including `npm update` within the
`_build/templates/default` directory.
2. Run `grunt build`. 
3. Clear your manager and browser cache, then browse around the manager
with your console open to verify all works as expected and no errors are
being reported.

Note that grunt build will spit out some warnings, as the versions of
bourbon and neat we need to stick with here (for now at least) are
ancient and contain some long-deprecated code. I attempted to bring
these dependencies up to date (including fontawesome) but there are many
breaking changes that make it difficult to unwind and get everything
working. Might try that later if there's enough "life" left in the 2.x
line and it's deemed beneficial to do so.

### Related issue(s)/PR(s)
Resolves issues with building after including the following PRs: #16493
and #16467.

---------

Co-authored-by: Jason Coward <jason@opengeek.com>