Log SQL errors during context cache generation (#16508)
### What does it do?
Checks for errors loading resources in the generation of the context
cache.

### Why is it needed?
This is used in the aliasMap and resourceMap, but would fail silently if
the query failed. In the specific case preceding this PR, the problem
was a borked migration, but it took too much sleuthing to figure out
where it was failing exactly due to this information not being
available.

### How to test
Make sure no message gets logged in normal use. 

### Related issue(s)/PR(s)
N/a