Add secondary regex option for transliterated file uploads (#16582)
### What does it do?
Adds a new system setting that allows transliterated uploads to override
the regex restriction used for file names (currently the same as with
Resource aliases).

### Why is it needed?
See the referenced issue. The request made there would be the most
common use-case, but I'd imagine there could be others where a user
wants the naming convention to be different between file names and
Resource aliases.

### How to test

1. Rebuild the core via `/_build/transport.core.php`
2. Run `/setup` (required to update the db based on the new build)
3. Test differing settings for Resources
(`friendly_alias_restrict_chars_pattern`) and the new setting for files
(`upload_translit_restrict_chars_pattern`). On a fresh install both
settings regex patterns will be the same, so be sure to change at least
one and verify the desired effect.

### Related issue(s)/PR(s)
Resolves #16551