Fix issues extending modSessionHandler and flushing all sessions (#16522)
### What does it do?

- Fixes the bugs from #15928, per #15934
- Refactored inititialization of the session handler and adopt the PHP
core SessionHandlerInterface, per #15934
- Move flushing sessions logic into the session handler to optionally
allow that to be extended, per #15957

### Why is it needed?

In two stale PRs #15928 and #15957 we have two proposals for dealing
with some bugs and extending session handlers. While trying to figure
out which one to use, I found both to have a solid approach, and wanting
to use one bit of one and another part of other PR.

This proposal replaces both those PRs as a middle ground, taking the
most benefit from both approachs.

### Breaking changes

This change includes a small breaking change in order to adopt the PHP
standard SessionHandlerInterface, namely adding parameters into the
open() method. There's not really any way around that in order to adopt
the standard.

As extending session handlers is a pretty advanced thing that I assume
not many people have done, I suggest we accept that for the sake of
getting closer to the standards, but make special note of this breaking
change in the docs for 3.1 and the announcement to make sure people
learn about it in case custom handlers need to be adjusted.

### How to test
Flush sessions

### Related issue(s)/PR(s)

Replaces #15934 and #15957
Fixes #15928

---------

Co-authored-by: sergant210 <sergant210@bk.ru>