Add Help! label to Menu Item (#16586)
Added Help! label to the Help Menu Item that was previously represented
by the just an icon.

This was to fix for #13485

### What does it do?
Provides accessible label for the Help button in the header #13485 

### Why is it needed?
The current link has no link text and no value for screen readers

### How to test
This may require a build and or rerunning setup. 

### Related issue(s)/PR(s)
Should resolve #13485 by adding aria-label and corresponding text via
lexicon to the `a` of the Help button in the header.