New password generation method: send user email (#16519)
### What does it do?

This is a re-up of #15461 originally by @sdrenth back in 2021, which has
gone stale waiting for some minor changes. I've rebased it, tweaked it,
tested it, so we can include it in 3.1.

This adds a new option for setting the password when creating/updating a
user: send the user a link to set their password. That's more secure and
builds upon improvements to the password reset flow that was done in
3.0.

### Why is it needed?

Showing the password on screen or manually setting a password is kinda
outdated and insecure.

### How to test

Create and/or edit a user, and choose "Let the user choose their own
password via email" for the password method. Look for the email (make
sure you have email delivery set up beforehand) and attempt to set the
new password.

### Related issue(s)/PR(s)

This PR replaces the stale PR #15461

https://github.com/Sterc/revolution/issues/22
https://github.com/modxcms/revolution/issues/13973
https://github.com/Sterc/revolution/pull/31

Co-authored-by: sander <sander@sterc.nl>