Add an asterisk to label of required TVs (#16594)
### What does it do?
Adds asterix automatically to a TV that has `allowBlank: false` set.


![tv_req_1](https://github.com/user-attachments/assets/5f9029ca-b076-4f56-b545-c40ffcd3fa32)


![tv_req_2](https://github.com/user-attachments/assets/296eac3a-d0ea-4ace-a336-284dfdd42f87)

### Why is it needed?
Required TVs are no different from regular ones. Only when saving the
resource, required TVs are marked with a message, **which is
inconvenient**.

p.s. For the 3.x branch, this behavior already exists, and it would not
be superfluous for 2.x as well.

### Related issue(s)/PR(s)
https://github.com/modxcms/revolution/issues/16593
https://github.com/modxcms/revolution/pull/14522