Fix visibility of the Install button in packages grid (#16675)
### What does it do?
Adds a new prop to provide a boolean indicator of a package's installed
state. The prop utilizes the recently-introduced date formatter's
`isEmpty` method, which more comprehensively determines whether the
`installed` field contains a meaningful value.

Note that this new prop still relies on a derived value. It would be
much better IMO at some future point to persist this state in a new
database field rather than basing it off of a date value.

### Why is it needed?
Install button is currently missing in the grid.

### How to test
Download, install, and uninstall a few packages to verify the install
button appears as expected.

### Related issue(s)/PR(s)
Resolves #16672