Convert invalid dates to NULL in modTransportPackage.installed (#16721)
### What does it do?
Converts any invalid dates in the modTransportPackage.installed column
to NULL.

### Why is it needed?
Empty dates should have been saved as NULL in the column and their
presence can cause errors when NO_ZERO_DATE/NO_ZERO_IN_DATE sql_modes
are enabled in an environment.

### How to test
In an environment with NO_ZERO_DATE/NO_ZERO_DATE sql_modes enabled,
build the transport and run the upgrade process, confirming that the
process completes and that any invalid dates in the
modTransportPackage.installed column are now NULL. You will need to have
invalid dates in that column before enabling the strict sql_modes if you
want to verify that this works properly.

### Related issue(s)/PR(s)
Required for #16717 to be merged.