Fix empty attributes when special characters appear in package readme (#16713)
### What does it do?
Change htmlentities to htmlspecialchars

### Why is it needed?
In one of my packages I have used a `ß` in the readme for the transport
package. When this transport package is installed, the result the of
`Workspace/Packages/GetAttribute` processor is empty for some reason
inside of the Parsedown class in combination with the processor response
and the package can't be installed anymore.

### How to test
Add a `ß` in the readme of a transport package.

### Related issue(s)/PR(s)
None known