Fix Template Access grid being initially empty when creating a new TV (#16678)
### What does it do?
Adds base grid method to differentiate the intent of the category URL
param, between being the value of a grid filter and that of the target
category for a new Element (when choosing _Create a [Element] Here_ via
the contextual menu), and applies the appropriate value for each
context.

### Why is it needed?
The Template Access grid initially is empty when a new Element is
created in the root (i.e., it has no category), because the value of the
category param is 'undefined'.

### How to test
1. Create new Elements of each type to verify normal operation.
2. Pay specific attention to the Template and TV Elements, where each
has an assignment grid (Template Variables and Template Access tabs
respectively). Verify these grids show as expected and are unfiltered
when creating a new Element and when editing an existing one.
3. On an existing Template and TV, go to the respective assignment grid
tab and use its category filter. Copy and paste the resulting URL (in
the browser URL bar) and copy that into a new browser tab. Verify that
the view in this new tab is the same as the previous tab where you
applied the filter. (This is to ensure the somewhat new feature of grid
filtering persistence via the URL still operates correctly.)

### Related issue(s)/PR(s)
Resolves #16677

---------

Co-authored-by: Jason Coward <jason@opengeek.com>