Avoid fatal error when Rest request fails in Package Browser (#16738)
### What does it do?
Better error handling in the `Workspace/Packages/Rest/GetList`
processor.

### Why is it needed?
While searching for extras in the "Package Browser", the code throws a
fatal error when there is a problem with the connection.

```
Uncaught TypeError: count(): Argument #1 ($value) must be of type Countable|array, string given
```

---

The `find()` function of the class `modTransportProvider` returns a
string if there is a error:


https://github.com/modxcms/revolution/blob/11d96669133bab7f1095c96be63368f841d8dbd7/core/src/Revolution/Transport/modTransportProvider.php#L333