Fix the annoying z-index issue in MODX 3.x
The reason for this issue can be found in https://coder-coder.com/z-index-isnt-working/#3-youre-setting-opacity-transform-or-other-css-properties-that-put-the-element-in-a-new-stacking-context

The flyouts are hidden with opacity: 0. This will put these elements into its own, new stacking context and the stacking context if the #modx-content will be fine with a z-index: 0.

Hopefully this will solve other z-index issues too.

modx-header .modx-subnav, #modx-footer .modx-subnav { z-index: 99999999999 } can be removed this way.