Fix z-index issue in the menu (#16778)
### What does it do?
Add a z-index: 0 to #modx-content and remove some not needed z-index in
the navbar.

### Why is it needed?
Fix the z-index issue in the menu of MODX 3.x. 

The reason for this issue can be found in
https://coder-coder.com/z-index-isnt-working/#3-youre-setting-opacity-transform-or-other-css-properties-that-put-the-element-in-a-new-stacking-context

The flyouts of the menu are hidden with opacity: 0. This will put these
elements into its own, new stacking context, which does not play well
with #modx-content items having a z-index. This will be fine when
#modx-content has a z-index: 0.

Hopefully this will solve other z-index issues too.

### How to test

Compile the CSS and check custom manager pages using elements with a
z-index (i.e. in dbAdmin, when Ace is installed).

### Related issue(s)/PR(s)
#16711 and others

#16690 and #16712 can be maybe reverted.

---------

Co-authored-by: Jason Coward <jason@opengeek.com>