Bump minimum PHP requirement to 8.1 (#16802)
### What does it do?
Bumps the minimum PHP requirement to the 8.1 release.

### Why is it needed?
A security vulnerability in the AWS-SDK-PHP dependency was fixed only in
releases that require PHP >= 8.1 which prevents installation of
dependencies via composer without ignorning vulnerabilities. I believe
the best course of action is dropping support for PHP < 8.1.

### How to test
Make sure everything still works with all of the dependency updates that
will come with bumping our minimum PHP to >= 8.1.

### Related issue(s)/PR(s)
N/A